[bugfix][ChromXControlEngine][Submodule] fixed MicroPIDValue type to quint32
[what] 1、修复Micro PID的值类型为quint32

[why] 之前写错了

[how]

+++ include/CCE_CommunicatEngine/CCESingleStatusPackage.h
@@ -95,7 +95,7 @@ class CCE_COMMUNICATENGINE_EXPORT CCESingleStatusPackage_ReadMicroPIDValue : pub
 {
     CCE_DECLARE_PACKAGECONSTRUCTOR(CCESingleStatusPackage_ReadMicroPIDValue, CCEAbstractSingleStatusPackage)
 public:
-    quint16 getValue() const;
+    quint32 getValue() const;
 protected:
     EFrameType CmdFrameType () const override {
         return EFrameType::EFT_ReadFrame;
